+++ gulpfile.js
@@ -3,7 +3,7 @@ const gulp = require('gulp');
 const mocha = require('gulp-spawn-mocha');
 
 const doctor = require('./src/tasks/doctor');
-
+const badmdl = require('./src/tasks/bad-mdl');
 const _ = require('lodash');
 const fs = require('fs-extra');
 
@@ -17,7 +17,8 @@ gulp.task('test', () =>
     )
 );
 
-gulp.task('doctor', callback => doctor({}, callback));
+gulp.task('bad-mdl', callback => badmdl({}, callback));
+gulp.task('doctor', ['bad-mdl'], callback => doctor({}, callback));
 
 gulp.task('updateREADME', () => {
     let README = fs.readFileSync('./README.md', 'utf8');